feat: finalize offline demo performance and accessible UI
Add bounded asynchronous SCHP keyframe inference with optical-flow propagation, deterministic color-stage optimizations, and explicit runtime telemetry.

Add shared Product/Diagnostic presentation shells, high-contrast light/dark themes, Unicode rendering, dependency locks, documentation, and model-independent regression coverage.

+++ src/chromalens/white_balance.py
@@ -210,8 +210,8 @@ class GrayWorldWhiteBalancer:
         )
 
         corrected_bgr = np.clip(
-            original_bgr.astype(np.float64)
-            * applied_gains.reshape((1, 1, 3)),
+            original_bgr.astype(np.float32)
+            * applied_gains.astype(np.float32).reshape((1, 1, 3)),
             0.0,
             255.0,
         ).astype(np.uint8)
